+++ packages/openapi3-parser/CHANGELOG.md
@@ -1,6 +1,6 @@
-# Fury OAS3 Parser Changelog
+# API Elements: OpenAPI 3 Parser Changelog
 
-## Master
+## 0.14.0 (2020-06-24)
 
 ### Enhancements